Remove OPUNet prefix from filenames
Remove abbreviations for GameSelectWindow filename
