+++ client/NetFixClient.vcxproj
@@ -118,16 +118,16 @@
     <ClCompile Include="FileSystemHelper.cpp" />
     <ClCompile Include="Log.cpp" />
     <ClCompile Include="Main.cpp" />
-    <ClCompile Include="OPUNetGameSelectWnd.cpp" />
-    <ClCompile Include="OPUNetTransportLayer.cpp" />
+    <ClCompile Include="GameSelectWindow.cpp" />
+    <ClCompile Include="TransportLayer.cpp" />
     <ClCompile Include="ValidatePacket.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="FileSystemHelper.h" />
     <ClInclude Include="Log.h" />
-    <ClInclude Include="OPUNetGameProtocol.h" />
-    <ClInclude Include="OPUNetGameSelectWnd.h" />
-    <ClInclude Include="OPUNetTransportLayer.h" />
+    <ClInclude Include="GameProtocol.h" />
+    <ClInclude Include="GameSelectWindow.h" />
+    <ClInclude Include="TransportLayer.h" />
     <ClInclude Include="resource.h" />
   </ItemGroup>
   <ItemGroup>
