+++ client/NetFixClient.vcxproj.filters
@@ -1,22 +1,22 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ResourceCompile Include="CustomDialogScript.rc"/>
+    <ResourceCompile Include="CustomDialogScript.rc" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="Main.cpp"/>
-    <ClCompile Include="OPUNetGameSelectWnd.cpp"/>
-    <ClCompile Include="OPUNetTransportLayer.cpp"/>
-    <ClCompile Include="ValidatePacket.cpp"/>
-    <ClCompile Include="Log.cpp"/>
-    <ClCompile Include="FileSystemHelper.cpp"/>
+    <ClCompile Include="Main.cpp" />
+    <ClCompile Include="ValidatePacket.cpp" />
+    <ClCompile Include="Log.cpp" />
+    <ClCompile Include="FileSystemHelper.cpp" />
+    <ClCompile Include="GameSelectWindow.cpp" />
+    <ClCompile Include="TransportLayer.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="OPUNetGameProtocol.h"/>
-    <ClInclude Include="OPUNetGameSelectWnd.h"/>
-    <ClInclude Include="OPUNetTransportLayer.h"/>
-    <ClInclude Include="Log.h"/>
-    <ClInclude Include="resource.h"/>
-    <ClInclude Include="FileSystemHelper.h"/>
+    <ClInclude Include="Log.h" />
+    <ClInclude Include="resource.h" />
+    <ClInclude Include="FileSystemHelper.h" />
+    <ClInclude Include="GameProtocol.h" />
+    <ClInclude Include="GameSelectWindow.h" />
+    <ClInclude Include="TransportLayer.h" />
   </ItemGroup>
 </Project>
